broke out passport functions to passport.js from app.js
